+++ src/main/java/ru/pelmegov/javashop/api/service/NewsService.java
@@ -10,7 +10,7 @@ public interface NewsService {
 
     Set<News> getNews(Integer count);
 
-    News getNewsById(Long id);
+    News getNewsById(Integer id);
 
     News getNewsByDate(Date date);
 }
